Rename UserDetails TeamSection to TeamsSection
