+++ graylog2-web-interface/src/components/users/UserDetails/TeamsSection.jsx
